requirements.txt
bleach==3.2.1
Flask==1.1.2
Flask-Avatars==0.2.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-HTTPAuth==4.1.0
Flask-Moment==0.10.0
Flask-Migrate==2.5.3
Flask-PageDown==0.3.0
Flask-SQLAlchemy==2.4.4
Flask-Session==0.3.2
Flask-WTF==0.14.3
email-validator==1.1.1
Markdown==3.3.3
psycopg2==2.8.6
requests==2.25.0
Werkzeug==1.0.1
WTForms-SQLAlchemy==0.2
